+++ docs/generate/generate.md
@@ -8,16 +8,16 @@ The API classes in this package provide a simple interface to read one or an ent
 
 For a single RAML file:
 ```javascript
-import { Api } from "@commerce-apps/raml-toolkit";
+import { generate } from "@commerce-apps/raml-toolkit";
 
-Api.init("path/to/my-spec.raml").then(console.log);
+generate.Api.init("path/to/my-spec.raml").then(console.log);
 ```
 
 For a group of RAML files:
 ```javascript
-import { ApiGroup } from "@commerce-apps/raml-toolkit";
+import { generate } from "@commerce-apps/raml-toolkit";
 
-ApiGroup.init([
+generate.ApiGroup.init([
     "path/to/my-spec.raml",
     "path/to/my-other-spec.raml"
 ]).then(console.log);
@@ -25,9 +25,9 @@ ApiGroup.init([
 
 Finally, for a collection of grouped RAML files:
 ```javascript
-import { ApiCollection } from "@commerce-apps/raml-toolkit";
+import { generate } from "@commerce-apps/raml-toolkit";
 
-ApiGroup.init({
+generate.ApiGroup.init({
     "My First Group": [
         "path/to/my-spec.raml",
         "path/to/my-other-spec.raml"
@@ -42,15 +42,15 @@ ApiGroup.init({
 handlebarsAmfHelpers module provides a number of functions to help easily access information from an AMF model. It also exports a Handlebars environment with all the handlebars AMF helper functions already registered. All the [handlebars-helpers](https://www.npmjs.com/package/handlebars-helpers) functions are also available through the Handlebars environment.
 
 ### Usage
-Using the handlebars environment exported by handlbarsAmfHelpers:
+Using the handlebars environment exported by generate namespace:
 ```javascript
 import {
-  handlebarsAmfHelpers
+  generate
 } from "@commerce-apps/raml-toolkit";
-const Handlebars = handlebarsAmfHelpers.handlebars;
+const Handlebars = generate.HandlebarsWithAmfHelpers;
 
 const template = Handlebars.compile(
-  fs.readFileSync(handlebarsTemplate, "utf8")
+  fs.readFileSync("path/to/my-template.ts.hbs", "utf8")
 );
 ```
