+++ src/index.ts
@@ -4,23 +4,10 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-
-export { getBearer } from "./download/bearerToken";
-export { RestApi } from "./download/exchangeTypes";
-export { ramlToolLogger } from "./common/logger";
-
-import * as amf from "amf-client-js";
-export { amf };
-export { model } from "amf-client-js";
-
-export * from "./generate";
-
+export * as amf from "amf-client-js";
+export * as amfParser from "./common/parser";
 export * as diff from "./diff";
-
-export {
-  getAllDataTypes,
-  getApiName,
-  getNormalizedName,
-  parseRamlFile,
-  resolveApiModel,
-} from "./common/parser";
+export * as download from "./download";
+export * as generate from "./generate";
+export * as lint from "./lint";
+export { ramlToolLogger } from "./common/logger";
