+++ src/lint/lint.ts
@@ -6,10 +6,18 @@
  */
 import path from "path";
 import { client, model, Core, MessageStyles, ProfileName } from "amf-client-js";
-
 import { parseRamlFile } from "../common/parser";
-import { profilePath } from ".";
 
+export const PROFILE_PATH = path.join(
+  __dirname,
+  "../../resources/lint/profiles"
+);
+
+/**
+ * Validate AMF model with the given profile
+ * @param amfModel - AMF model
+ * @param profileFile - Path to the profile file
+ */
 export async function validateCustom(
   amfModel: model.document.BaseUnit,
   profileFile: string
@@ -30,18 +38,28 @@ export async function validateCustom(
   return report;
 }
 
+/**
+ * Validate AMF model with the given profile
+ * @param model - AMF model
+ * @param profile - Name of the profile
+ */
 export async function validateModel(
   model: model.document.BaseUnit,
   profile: string
 ): Promise<client.validate.ValidationReport> {
   const results = await validateCustom(
     model,
-    `file://${path.join(profilePath, `${profile}.raml`)}`
+    `file://${path.join(PROFILE_PATH, `${profile}.raml`)}`
   );
 
   return results;
 }
 
+/**
+ * Print validation results to console
+ * @param results - Validation results
+ * @param warnings - True to print warnings
+ */
 export async function printResults(
   results: client.validate.ValidationReport,
   warnings = false
@@ -58,6 +76,11 @@ Number of hidden warnings: ${results.results.length}
   }
 }
 
+/**
+ * Validate API specification file with the given profile
+ * @param filename - Path to the API specification file
+ * @param profile - Name of the profile
+ */
 export async function validateFile(
   filename: string,
   profile: string
