+++ src/lint/lintCommand.test.ts
